+++ lydian/cogs/debug.py
@@ -13,32 +13,27 @@ def debug_enabled(_ctx: commands.Context) -> bool:
 class DebugCog(commands.Cog):
     """Debugging or testing commands for development purposes."""
 
-    @commands.command()
-    @commands.check(debug_enabled)
+    @commands.command(checks=[debug_enabled])
     async def sendinfo(self, ctx: commands.Context) -> None:
         """Send a test message using :py:func:`lydian.cogs.util.embed_info` to make the embed."""
         await ctx.send(embed=embed_info('Information', 'Some general event has occurred.'))
 
-    @commands.command()
-    @commands.check(debug_enabled)
+    @commands.command(checks=[debug_enabled])
     async def sendok(self, ctx: commands.Context) -> None:
         """Send a test message using :py:func:`lydian.cogs.util.embed_ok` to make the embed."""
         await ctx.send(embed=embed_ok('Success', 'Everything is OK.'))
 
-    @commands.command()
-    @commands.check(debug_enabled)
+    @commands.command(checks=[debug_enabled])
     async def sendwarn(self, ctx: commands.Context) -> None:
         """Send a test message using :py:func:`lydian.cogs.util.embed_warn` to make the embed."""
         await ctx.send(embed=embed_warn('Warning', 'Something unexpected happened.'))
 
-    @commands.command()
-    @commands.check(debug_enabled)
+    @commands.command(checks=[debug_enabled])
     async def senderror(self, ctx: commands.Context) -> None:
         """Send a test message using :py:func:`lydian.cogs.util.embed_error` to make the embed."""
         await ctx.send(embed=embed_error('Error', 'Something failed.'))
 
-    @commands.command()
-    @commands.check(debug_enabled)
+    @commands.command(checks=[debug_enabled])
     async def fail(self, _ctx: commands.Context) -> None:
         """Raises a ``ValueError``."""
         logger.debug('About to raise ValueError...')