+++ lydian/cogs/general.py
@@ -2,6 +2,7 @@
 from discord.ext import commands
 
 from lydian.cogs.util import embed_info
+from lydian.config import config
 
 
 class GeneralCog(commands.Cog):
@@ -10,7 +11,7 @@ class GeneralCog(commands.Cog):
     def __init__(self, bot: commands.bot.Bot) -> None:
         self.bot = bot
 
-    @commands.command()
+    @commands.command(aliases=config.command_aliases.get('hello', ()))
     async def hello(self, ctx: commands.Context) -> None:
         """Test command."""
         await ctx.send(embed=embed_info('Hello, world!'))